--- src/test/java/kitchenpos/ControllerTest.java
@@ -1,32 +1,0 @@
-package kitchenpos;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.junit.jupiter.api.BeforeEach;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.data.jpa.mapping.JpaMetamodelMappingContext;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-
-@MockBean(JpaMetamodelMappingContext.class)
-public class ControllerTest {
-    @Autowired
-    protected MockMvc webMvc;
-
-    @Autowired
-    private WebApplicationContext context;
-
-    protected ObjectMapper objectMapper = new ObjectMapper();
-
-    @BeforeEach
-    public void setUp() {
-        this.webMvc = MockMvcBuilders.webAppContextSetup(context)
-                .addFilters(new CharacterEncodingFilter("UTF-8", true))
-                .alwaysDo(print())
-                .build();
-    }
-}
